requirements.txt
# Time-Series Forecasting and Anomaly Detection
# Production ML Repository

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Model persistence
joblib>=1.3.0

# Optional: for notebook support
jupyter>=1.0.0
ipykernel>=6.25.0
